Experiment_Example/3_XRay_experiement/CCD_Xray/240927_autom_scans/requirements.txt
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astkit==0.5.4
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
broadbean==0.14.0
certifi==2024.8.30
cf_xarray==0.9.1
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cma==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.6.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.6.0
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==7.1.0
instrumental==0.5.3
ipykernel==6.29.4
ipympl==0.9.4
ipython==8.25.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
libximc==2.14.23
llvmlite==0.42.0
locket==1.0.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
opencv-python==4.9.0.80
opentelemetry-api==1.25.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pillow==10.3.0
pip3-autoremove==1.2.2
platformdirs==4.2.2
plumbum==1.8.3
prometheus_client==0.21.0
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
pyaml==24.7.0
pyarrow==16.1.0
pycparser==2.22
Pygments==2.18.0
pylablib==1.4.3
pylablib-lightweight==1.4.3
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
pyserial==3.5
pytango==9.5.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyusb==1.2.1
PyVISA==1.14.1
PyVISA-py==0.7.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qcodes==0.45.0
qcodes_contrib_drivers==0.22.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rpyc==6.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
schema==0.7.7
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.13.1
Send2Trash==1.8.3
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uncertainties==3.2.1
uri-template==1.3.0
urllib3==2.2.3
versioningit==3.1.1
virtualenv==20.26.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xarray==2024.6.0
zipp==3.19.2
